refacto: directory done

+++ incs/packet_capture/checksum.h
@@ -5,7 +5,7 @@
 #include <netinet/tcp.h>
 #include <netinet/udp.h>
 
-uint16_t ip_checksum(uint16_t *buf, int nwords);
-uint16_t tcp_checksum(uint16_t *ptr, int nbytes, uint64_t src_addr, uint64_t dest_addr);
+uint16_t ip_checksum(uint8_t *buf, int nwords);
+uint16_t tcp_checksum(uint8_t *ptr, int nbytes, uint64_t src_addr, uint64_t dest_addr);
 
 #endif // CHECKSUM_H
